+++ packages/core/src/index.ts
@@ -1,0 +1,5 @@
+import { createDrive } from './drive';
+
+export { Drive } from './drive';
+export { Films, Shows } from './types/content';
+export default createDrive;