+++ packages/server/src/auth/users.ts
@@ -1,0 +1,11 @@
+export default class Users {
+  private users: Map<string, string> = new Map();
+
+  upsert(username: string, password: string) {
+    this.users.set(username, password);
+  }
+
+  get(username: string) {
+    return this.users.get(username);
+  }
+}