+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,43 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.9](https://github.com/getholo/stream/compare/@getholo/stream@1.0.0-alpha.8...@getholo/stream@1.0.0-alpha.9) (2019-11-22)
+
+
+### Bug Fixes
+
+* tv shows not working with kodi ([faee79d](https://github.com/getholo/stream/commit/faee79d441c4c819ad0d84d717eafa17c0a63a6c))
+
+
+
+
+
+# [1.0.0-alpha.8](https://github.com/getholo/stream/compare/@getholo/stream@1.0.0-alpha.6...@getholo/stream@1.0.0-alpha.8) (2019-11-12)
+
+
+### Bug Fixes
+
+* cli was missing bin ([3ea1c86](https://github.com/getholo/stream/commit/3ea1c86510793a8fda2fb310aecc457c08e1cbfa))
+
+
+
+
+
+# 1.0.0-alpha.7 (2019-11-12)
+
+
+### Features
+
+* Add support for custom folder structures ([#5](https://github.com/getholo/stream/issues/5)) ([26905ed](https://github.com/getholo/stream/commit/26905ed8b7ad9f56bf694628d913e4863ef95680))
+* Allow custom password to be set ([#3](https://github.com/getholo/stream/issues/3)) ([0bb3195](https://github.com/getholo/stream/commit/0bb3195bdb3f641e05c7b8013365729b36625647))
+* Core Components ([86bf5b6](https://github.com/getholo/stream/commit/86bf5b62d08181947ab453f7ee302ae417b77f58))
+* fetch changes in Drive periodically ([#6](https://github.com/getholo/stream/issues/6)) ([a668c15](https://github.com/getholo/stream/commit/a668c152c2151d437355022ce89c6fc8e4cbf83c))
+* TV Shows support ([#10](https://github.com/getholo/stream/issues/10)) ([1a650cf](https://github.com/getholo/stream/commit/1a650cf48b802e5713f973908dfa5e30811a34b8))
+
+
+
+
+
 # [1.0.0-alpha.6](https://github.com/getholo/stream/compare/@getholo/stream@1.0.0-alpha.5...@getholo/stream@1.0.0-alpha.6) (2019-11-01)
 
 **Note:** Version bump only for package @getholo/stream