+++ packages/server/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-alpha.8](https://github.com/getholo/stream/compare/@getholo/stream-server@1.0.0-alpha.7...@getholo/stream-server@1.0.0-alpha.8) (2019-11-22)
+
+
+### Bug Fixes
+
+* tv shows not working with kodi ([faee79d](https://github.com/getholo/stream/commit/faee79d441c4c819ad0d84d717eafa17c0a63a6c))
+
+
+
+
+
+# [1.0.0-alpha.7](https://github.com/getholo/stream/compare/@getholo/stream-server@1.0.0-alpha.6...@getholo/stream-server@1.0.0-alpha.7) (2019-11-12)
+
+
+### Features
+
+* add last modified date ([#12](https://github.com/getholo/stream/issues/12)) ([11ba52e](https://github.com/getholo/stream/commit/11ba52e908bfce30b00bfa42a4a338ab7cb70c76))
+
+
+
+
+
 # [1.0.0-alpha.6](https://github.com/getholo/stream/compare/@getholo/stream-server@1.0.0-alpha.5...@getholo/stream-server@1.0.0-alpha.6) (2019-11-01)